info about me
Me? I am the same character as you trying to live out this new storyline. 
God knows if it'll go anywhere or work. 
But i WOULDNT be doing it if i didn't think there was a chance, now would i?
I, like you, think that I am special in some way. How that manifests for me is in my belief that I have access to some sort of creative space of abstract ideas that nerdy coder types don't. 
This may come to be true, but more likely may be a useful story that I tell myself which enables me to work with confidence, get better, and eventually emerge and contribute to the space in a 'special' and unique manner. 
Then I will have been right all along, even if i wasn't. and even if i never said it.